Move method to create watch channels requests to logic
